Fix an inconsistency with member creation
Previously, adding a member would create a row in the users_yearly
table.  This would normally have been fine, however, it is assumed in
access level computation for login that any user with a row in the table
has member level access.  This is something we might want to revisit
later and find a better solution as not adding a row takes away the
ability to check when a member first signed up (a timestamp might not be
the worst idea for that either)