add email to be sent some period of time before projects
I added a cron page which is meant to be called by cron every so often.
Right now, the only task accomplished by the page is to send a reminder
email for project some amount of time (2 days by default) before a
project.  Other tasks can be added to the page later.